Switch the script to a different model
Now, instead of having 1 big playbook with all and using limit,
we will have multiple playbook, and the script will be smart into
deploying what need to be deployed. This permit to group everything
more logically (like all nagios in 1 playbook), and thus
permit to deploy something on all servers without deploying everything
on all servers again.

Next step is to get the best of both world, and keep the capacity
to use -l if a role affect a subpart of the playbook, to ease migration.