Add more verification, and use ssh rather than keyscan
While the hostname is supposed to be correct (since that's
only for authorized users, and with git), I prefer to be extra
cautious and check arguments passed to shell. I also found out
that pipe redirection would likely not work, so I had
to find a work around using ssh (no need for real connect however,
so that's why I do not care about the username).