Add a cache for parse_roles_meta
A run on the whole repository show that this function is called
several time, parsing endlessly the same yaml file. While a cleaner solution
would be to use a fancy memoization decorator, i try to reduce the requirements,
and so rather do it by hand.

The test I did show that I can reduce a run of 3m to 15s for a repo
with 474 commits and 12m of source. That's still too much but acceptable
enough I guess.