Use safe_load, to avoid potential security issue
While only trusted users can commit/push, and most of them would
already be admins on the infrastructure, there is a undocumented
experimental mode permitting to use git_shell.

If this mode was used, then a user would be able to push code
that would result in code execution and bypass the git-shell
restriction. It may also result in leaving not much trace, since
this hook may prevent commits, and so a attacker could push a commit
resulting in a code execution and deny the push from happening.

A CVE was requested and this commit message will be edited to reflect
the result.