Replace the ansible_admin group by a committers group
This also deprecated the allow_ansible_commands, and replace it
with a group based access. The main driver of the change is to
be able to use gerrit for doing code review, and so be able
to give more privileges to humans admins (such as me) than
to a potential gerrit bot account that would be only able
to push and not run anything else. The idea being that in
case a of a gerrit compromision, the attacker will not be
able to do anything that would end unaudited (such as running
a custom ansible script).

This also clarify exactly the level of access given to each users,
and provide a 2 tier ACL system permitting to ease integration
of new comers.