Fix potential code execution due to VariableManager plugin loading
While investigating a problem on the code, I found out that ansible
do load plugin from the directory where the hosts file is when
we use VariableManager.

So since the file used is stored in /tmp, this mean that ansible
will try to load plugins from /tmp/vars_plugin, resulting in potential
code execution as a different user when triggering a git commit.

This would matter only in very specific setup (since there is already
easier path to deploy source code, that's the whole point of the role
itself), and requires a existing user to have access to /tmp, hence to the
server.

One role that can be used to mitigate this and later /tmp issue is
https://github.com/OSAS/ansible-role-polyinst_tmp