Fix the OSError exception seen since a few days
When adding a new hosts, a error message do appear:

    Exception OSError: (2, 'No such file or directory', '/tmp/tmp7Cq_Us/tmpxXIavL') in <bound method _TemporaryFileWrapper.__del__ of <closed file '<fdopen>', mode 'w+' at 0x2cfd6f0>> ignored

It turn out to be caused by a mismatch of the object lifecycle. When tmp_file
is out of scope, it get erased, but the file is erased before by the rmtree
function, hence the exception. While this doesn't affect functionnality, this
is a bit ugly and misleading, so better fix it.