Fix the bastion lib to work with 2.4
Due to the change regarding inventory in 2.4, the code to extract
variable from inventory has changed a lot. However, since ansible 2.3
is now in RHEL, we have to keep compatibility with it for potentially
a long time.

So this code is a attempt at that, using a wrapper class to abstract
away the differences.

Once 2.3 is no longer supported (and no longer shipped), the code
will be cleaned again and fully ported to 2.4 only.