Add a link, to handle update_galaxy.sh running as root
Since it run as root, collection are installed as root. But
ansible is run as a unprivileged user, so it can't access the newly
installed collection.

There is several way to fix this:
- running ansible galaxy as non root would mean some duplication, or
  using sudo/su, which bring complication
- using a config file and a custom path requires to have the script
aware of the uid to us, not optimal
- installing in /usr/share is prone to clash with rpms

So a link is the less hacky way to do it for now (but I may change my
mind later, cause that's still a hack)